+++ activejob/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Rails 5.1.0.alpha ##
 
+*   Remove deprecated ability to pass an adapter class to `ActiveJob::QueueAdapter#queue_adapter=`.
+
+    *Jon Moss*
+
 *   Remove deprecated `original_exception` method from `ActiveJob::DeserializationError`.
 
     *Jon Moss*